web-animations/interfaces/KeyframeEffect/keyframe-properties.html

<!DOCTYPE html>
<meta charset='utf-8'>
<title>Web Animations API: Keyframe Property tests</title>
<link rel='help' href='https://w3c.github.io/web-animations/#processing-a-keyframes-argument'>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../../testcommon.js'></script>
<script src='../../resources/keyframe-utils.js'></script>
<body>
<script>
'use strict';

test(function(t) {
  var keyframe = {};
  Object.defineProperty(keyframe, 'width', {value: '200px'});
  Object.defineProperty(keyframe, 'height', {
    value: '100px',
    enumerable: true});
  assert_equals(keyframe.width, '200px', 'width of keyframe is readable');
  assert_equals(keyframe.height, '100px', 'height of keyframe is readable');
  var anim = createDiv(t).animate([keyframe, {height: '200px'}], 1);
  assert_frame_lists_equal(anim.effect.getKeyframes(), [
      {offset: null, computedOffset: 0, easing: 'linear', height: '100px'},
      {offset: null, computedOffset: 1, easing: 'linear', height: '200px'},
  ]);
}, 'Only enumerable properties on keyframes are considered');

test(function(t) {
  var KeyframeParent = function() { this.width = '100px'; };
  KeyframeParent.prototype = { height: '100px' };
  var Keyframe = function() { this.top = '100px'; };
  Keyframe.prototype = Object.create(KeyframeParent.prototype);
  Object.defineProperty(Keyframe.prototype, 'left', {
    value: '100px',
    enumerable: 'true'});
  var keyframe = new Keyframe();
  var anim = createDiv(t).animate([keyframe, {top: '200px'}], 1);
  assert_frame_lists_equal(anim.effect.getKeyframes(), [
      {offset: null, computedOffset: 0, easing: 'linear', top: '100px'},
      {offset: null, computedOffset: 1, easing: 'linear', top: '200px'},
  ]);
}, 'Only properties defined directly on keyframes are considered');

</script>
</body>
